+++ docs/documentation/docs/about/release-notes.md
@@ -5,6 +5,7 @@
 **New controls**
 
 - `PropertyFieldMultiSelect` got added
+- `PropertyFieldNumber` got added
 
 **Enhancements**